+++ frontend/vite.config.ts
@@ -1,7 +1,12 @@
-import {defineConfig} from 'vite'
-import {svelte} from '@sveltejs/vite-plugin-svelte'
+import { svelte } from '@sveltejs/vite-plugin-svelte';
+import { defineConfig } from 'vite';
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [svelte()]
+  plugins: [svelte()],
+  resolve: {
+    alias: {
+      "@api": __dirname + "/wailsjs/go/ui"
+    }
+  }
 })
