+++ main.go
@@ -31,6 +31,8 @@ func Launch() {
 
 	// Create an instance of the app structure
 	app := ui.NewApp()
+	k8s := ui.NewK8s()
+	pkg := ui.NewPackage()
 
 	// Create application with options
 	err := wails.Run(&options.App{
@@ -41,7 +43,8 @@ func Launch() {
 		BackgroundColour: &options.RGBA{R: 27, G: 38, B: 54, A: 1},
 		OnStartup:        app.Startup,
 		Bind: []interface{}{
-			app,
+			k8s,
+			pkg,
 		},
 	})
