+++ src/ui/packages.go
@@ -1,0 +1,10 @@
+package ui
+
+// Package struct
+type Package struct {
+}
+
+// NewPackage creates a new Package application struct
+func NewPackage() *Package {
+	return &Package{}
+}
